messaging service big refactor
