+++ packages/bp/src/core/app/botpress.ts
@@ -436,7 +436,7 @@ export class Botpress {
         this.eventCollector.storeEvent(event)
       }
 
-      this.messagingService.informProcessingDone(event)
+      this.messagingService.collector.informProcessingDone(event)
       await converseApiEvents.emitAsync(`done.${buildUserKey(event.botId, event.target)}`, event)
     }
