+++ packages/bp/src/core/messaging/messaging-router.ts
@@ -14,7 +14,7 @@ export class MessagingRouter extends CustomRouter {
   }
 
   public setupRoutes() {
-    this.messagingService.messaging.setup(this.router, '/receive')
+    this.messagingService.interactor.client.setup(this.router, '/receive')
     this.legacy.setup()
   }
 }
